requirements.txt
tensorflow-gpu
numpy
scikit-learn
seaborn
matplotlib